#1 Adding RollingUpdates strategy and clean up
