+++ config/manager/manager.yaml
@@ -2,7 +2,7 @@ apiVersion: v1
 kind: ServiceAccount
 metadata:
   name: webpage-operator
-  namespace: default
+  namespace: local-ns
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
@@ -38,7 +38,7 @@ metadata:
   name: webpage-operator
   namespace: default
 spec:
-  replicas: 1
+  replicas: 2
   selector:
     matchLabels:
       app: webpage-operator
@@ -52,3 +52,22 @@ spec:
       - name: manager
         image: webpage-operator:latest
         imagePullPolicy: IfNotPresent
+  strategy:
+    type: RollingUpdate 
+    rollingUpdate:
+      maxSurge: 25%
+      maxUnavailable:  25% 
+
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: webpage-operator
+spec:
+  selector:
+    app: webpage-operator
+  ports:
+  - name: http
+    port: 80
+    targetPort: 8080
+  type: ClusterIP
